refactor(tiptap): lint fix everything (#448)
* docs(tiptap): update docs/tiptap and tiptap/README

* fix(tiptap): linter errors

* refactor(tiptap): handle more eslint errors

* feat(tiptap): memoize updateAttributes

* fix(tiptap): undo-redo button active

* fix(tiptap): rm disabled buttons hover style

+++ tiptap/src/components/tiptap-extensions/folio-tiptap-columns/folio-tiptap-columns-node.ts
@@ -66,17 +66,20 @@ export const FolioTiptapColumnsNode = Node.create({
           },
       addFolioTiptapColumnBefore:
         () =>
-          ({ dispatch, state }) => {
+          ({ dispatch, state }: CommandParams) => {
+            if (!dispatch) return false;
             return addOrDeleteColumn({ dispatch, state, type: 'addBefore' });
           },
       addFolioTiptapColumnAfter:
         () =>
-          ({ dispatch, state }) => {
+          ({ dispatch, state }: CommandParams) => {
+            if (!dispatch) return false;
             return addOrDeleteColumn({ dispatch, state, type: 'addAfter' });
           },
       deleteFolioTiptapColumn:
         () =>
-          ({ dispatch, state }) => {
+          ({ dispatch, state }: CommandParams) => {
+            if (!dispatch) return false;
             return addOrDeleteColumn({ dispatch, state, type: 'delete' });
           },
     };
